+++ client/src/www/app/outline_server_repository/config.spec.ts
@@ -15,139 +15,6 @@
 import * as config from './config';
 import * as method_channel from '../method_channel';
 
-/*
-describe('getAddressFromTransport', () => {
-  it('extracts address', () => {
-    expect(
-      config.TEST_ONLY.getAddressFromTransportConfig({
-        host: 'example.com',
-        port: 443,
-      })
-    ).toEqual({host: 'example.com', port: 443});
-    expect(
-      config.TEST_ONLY.getAddressFromTransportConfig({
-        host: '1:2::3',
-        port: 443,
-      })
-    ).toEqual({host: '1:2::3', port: 443});
-    expect(
-      config.TEST_ONLY.getAddressFromTransportConfig({host: 'example.com'})
-    ).toEqual({host: 'example.com', port: undefined});
-    expect(
-      config.TEST_ONLY.getAddressFromTransportConfig({host: '1:2::3'})
-    ).toEqual({host: '1:2::3', port: undefined});
-  });
-
-  it('fails on invalid config', () => {
-    expect(config.TEST_ONLY.getAddressFromTransportConfig({})).toBeUndefined();
-  });
-});
-
-describe('setTransportHost', () => {
-  it('sets host', () => {
-    expect(
-      JSON.stringify(
-        config.setTransportConfigHost(
-          {host: 'example.com', port: 443},
-          '1.2.3.4'
-        )
-      )
-    ).toEqual('{"host":"1.2.3.4","port":443}');
-    expect(
-      JSON.stringify(
-        config.setTransportConfigHost(
-          {host: 'example.com', port: 443},
-          '1:2::3'
-        )
-      )
-    ).toEqual('{"host":"1:2::3","port":443}');
-    expect(
-      JSON.stringify(
-        config.setTransportConfigHost({host: '1.2.3.4', port: 443}, '1:2::3')
-      )
-    ).toEqual('{"host":"1:2::3","port":443}');
-  });
-
-  it('fails on invalid config', () => {
-    expect(config.setTransportConfigHost({}, '1:2::3')).toBeUndefined();
-  });
-});
-
-describe('parseTunnelConfig', () => {
-  it('parses correctly', async () => {
-    expect(
-      await config.parseTunnelConfig(
-        '{"server": "example.com", "server_port": 443, "method": "METHOD", "password": "PASSWORD"}'
-      )
-    ).toEqual({
-      firstHop: 'example:com:443',
-      transport: `{
-        host: 'example.com',
-        port: 443,
-        method: 'METHOD',
-        password: 'PASSWORD',
-      }`,
-    });
-  });
-
-  it('parses prefix', async () => {
-    expect(
-      await config.parseTunnelConfig(
-        '{"server": "example.com", "server_port": 443, "method": "METHOD", "password": "PASSWORD", "prefix": "POST "}'
-      )
-    ).toEqual({
-      firstHop: 'example:com:443',
-      transport: `{
-        endpoint: 'example.com:443',
-        method: 'METHOD',
-        password: 'PASSWORD',
-        prefix: 'POST ',
-      }`,
-    });
-  });
-
-  it('parses URL', async () => {
-    const ssUrl = SIP002_URI.stringify(
-      makeConfig({
-        host: 'example.com',
-        port: 443,
-        method: 'chacha20-ietf-poly1305',
-        password: 'PASSWORD',
-      })
-    );
-    expect(await config.parseTunnelConfig(ssUrl)).toEqual({
-      firstHop: 'example:com:443',
-      transport: `{
-        host: 'example.com',
-        port: 443,
-        method: 'chacha20-ietf-poly1305',
-        password: 'PASSWORD',
-      }`,
-    });
-  });
-
-  it('parses URL with blanks', async () => {
-    const ssUrl = SIP002_URI.stringify(
-      makeConfig({
-        host: 'example.com',
-        port: 443,
-        method: 'chacha20-ietf-poly1305',
-        password: 'PASSWORD',
-      })
-    );
-    expect(await config.parseTunnelConfig(`  ${ssUrl} \n\n\n`)).toEqual({
-      firstHop: 'example:com:443',
-      transport: `{
-        host: 'example.com',
-        port: 443,
-        method: 'chacha20-ietf-poly1305',
-        password: 'PASSWORD',
-      }`,
-    });
-  });
-});
-*/
-
 describe('parseAccessKey', () => {
   method_channel.installDefaultMethodChannel({
     async invokeMethod(methodName: string, params: string): Promise<string> {
